BUG: Fixed heap corruption error by log handlers
Problem:
When a multi-threaded filter logged a message during execution it
often happened that multiple threads logged a message at the same time,
using the same message handler. The message handler contained a single
QRegExp ContextRegExp object, which was used from multiple threads
at the same time and caused memory corruption and crash.

Analysis:
Potential solutions could be: lock the QRegExp object so that only one
thread can use it at a time or create a new QRegExp for each thread.
Long-term nice solution would be to change ITK and VTK to provide
interface for structured logging (send filename, line number, log level,
message separately) instead of dumping a text message that we have to
parse.

Solution:
Creating a new QRegExp for each thread is simpler and probably
not slower, therefore this option is implemented.

Fixes #546

Reviewed-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>