ENH: 4236 Show completion behind function(). fixed
There was a bug linked with the autocompletion, for example :
if you type in command
import SampleData
SampleData.SampleDataLogic().dow... if you try to autocomplete with 'Tab' it won't work.
The autocompletion was working only if we removed the '()' but in command you must have the '()' to make the command works.
Then, we should leave the '()' for the command shell but removed it to search the autocompletion.